+++ src/Feature/Sms/SmsHttpFeature.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace Smsapi\Client\Feature\Sms;
 
+use Psr\Http\Client\ClientInterface;
 use Smsapi\Client\Feature\Data\DataFactoryProvider;
 use Smsapi\Client\Feature\Sms\Bag\DeleteScheduledSmssBag;
 use Smsapi\Client\Feature\Sms\Bag\ScheduleSmsBag;
@@ -25,13 +26,16 @@ use stdClass;
  */
 class SmsHttpFeature implements SmsFeature
 {
+    private $externalHttpClient;
     private $requestExecutorFactory;
     private $dataFactoryProvider;
 
     public function __construct(
+        ClientInterface $externalHttpClient,
         RequestExecutorFactory $requestExecutorFactory,
         DataFactoryProvider $dataFactoryProvider
     ) {
+        $this->externalHttpClient = $externalHttpClient;
         $this->requestExecutorFactory = $requestExecutorFactory;
         $this->dataFactoryProvider = $dataFactoryProvider;
     }
@@ -39,7 +43,7 @@ class SmsHttpFeature implements SmsFeature
     public function sendernameFeature(): SendernamesFeature
     {
         return new SendernamesHttpFeature(
-            $this->requestExecutorFactory->createRestRequestExecutor(),
+            $this->requestExecutorFactory->createRestRequestExecutor($this->externalHttpClient),
             $this->dataFactoryProvider->provideSendernameFactory()
         );
     }
@@ -216,6 +220,8 @@ class SmsHttpFeature implements SmsFeature
      */
     private function makeRequest($data): stdClass
     {
-        return $this->requestExecutorFactory->createLegacyRequestExecutor()->request('sms.do', (array)$data);
+        return $this->requestExecutorFactory
+            ->createLegacyRequestExecutor($this->externalHttpClient)
+            ->request('sms.do', (array)$data);
     }
 }