+++ src/Infrastructure/RequestExecutor/LegacyRequestExecutor.php
@@ -5,7 +5,9 @@ declare(strict_types=1);
 namespace Smsapi\Client\Infrastructure\RequestExecutor;
 
 use Psr\Http\Client\ClientInterface;
-use Smsapi\Client\Infrastructure\RequestAssembler\GuzzleRequestAssembler;
+use Psr\Http\Message\RequestFactoryInterface;
+use Psr\Http\Message\StreamFactoryInterface;
+use Smsapi\Client\Infrastructure\RequestAssembler\RequestAssembler;
 use Smsapi\Client\Infrastructure\RequestMapper\LegacyRequestMapper;
 use Smsapi\Client\Infrastructure\ResponseMapper\LegacyResponseMapper;
 use stdClass;
@@ -18,25 +20,28 @@ class LegacyRequestExecutor
     private $requestMapper;
     private $client;
     private $legacyResponseMapper;
-    private $requestAssembler;
+    private $requestFactory;
+    private $streamFactory;
 
     public function __construct(
         LegacyRequestMapper $requestMapper,
         ClientInterface $client,
         LegacyResponseMapper $legacyResponseMapper,
-        GuzzleRequestAssembler $requestAssembler
+        RequestFactoryInterface $requestFactory,
+        StreamFactoryInterface $streamFactory
     ) {
         $this->requestMapper = $requestMapper;
         $this->client = $client;
         $this->legacyResponseMapper = $legacyResponseMapper;
-        $this->requestAssembler = $requestAssembler;
+        $this->requestFactory = $requestFactory;
+        $this->streamFactory = $streamFactory;
     }
 
     public function request(string $path, array $builtInParameters, array $userParameters = []): stdClass
     {
         $request = $this->requestMapper->map($path, $builtInParameters, $userParameters);
 
-        $assembledRequest = $this->requestAssembler->assemble($request);
+        $assembledRequest = (new RequestAssembler($this->requestFactory, $this->streamFactory))->assemble($request);
 
         $response = $this->client->sendRequest($assembledRequest);