+++ src/Service/SmsapiPlHttpService.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Smsapi\Client\Service;
 
+use Psr\Http\Client\ClientInterface;
 use Smsapi\Client\Feature\Data\DataFactoryProvider;
 use Smsapi\Client\Feature\Mms\MmsFeature;
 use Smsapi\Client\Feature\Mms\MmsHttpFeature;
@@ -20,13 +21,16 @@ class SmsapiPlHttpService implements SmsapiPlService
 {
     use HttpDefaultFeatures;
 
+    private $externalHttpClient;
     private $requestExecutorFactory;
     private $dataFactoryProvider;
 
     public function __construct(
+        ClientInterface $externalHttpClient,
         RequestExecutorFactory $requestExecutorFactory,
         DataFactoryProvider $dataFactoryProvider
     ) {
+        $this->externalHttpClient = $externalHttpClient;
         $this->requestExecutorFactory = $requestExecutorFactory;
         $this->dataFactoryProvider = $dataFactoryProvider;
     }
@@ -34,7 +38,7 @@ class SmsapiPlHttpService implements SmsapiPlService
     public function mmsFeature(): MmsFeature
     {
         return new MmsHttpFeature(
-            $this->requestExecutorFactory->createLegacyRequestExecutor(),
+            $this->requestExecutorFactory->createLegacyRequestExecutor($this->externalHttpClient),
             $this->dataFactoryProvider->provideMmsFactory()
         );
     }
@@ -42,7 +46,7 @@ class SmsapiPlHttpService implements SmsapiPlService
     public function vmsFeature(): VmsFeature
     {
         return new VmsHttpFeature(
-            $this->requestExecutorFactory->createLegacyRequestExecutor(),
+            $this->requestExecutorFactory->createLegacyRequestExecutor($this->externalHttpClient),
             $this->dataFactoryProvider->provideVmsFactory()
         );
     }
@@ -50,7 +54,7 @@ class SmsapiPlHttpService implements SmsapiPlService
     public function profileFeature(): SmsapiPlProfileFeature
     {
         return new SmsapiPlProfileHttpFeature(
-            $this->requestExecutorFactory->createRestRequestExecutor(),
+            $this->requestExecutorFactory->createRestRequestExecutor($this->externalHttpClient),
             $this->dataFactoryProvider
         );
     }