+++ src/SmsapiHttpClient.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace Smsapi\Client;
 
+use Psr\Http\Client\ClientInterface;
+use Psr\Http\Message\RequestFactoryInterface;
+use Psr\Http\Message\StreamFactoryInterface;
 use Psr\Log\LoggerAwareTrait;
 use Psr\Log\NullLogger;
 use Smsapi\Client\Feature\Data\DataFactoryProvider;
-use Smsapi\Client\Infrastructure\Client\GuzzleClientFactory;
+use Smsapi\Client\Infrastructure\HttpClient\HttpClientFactory;
 use Smsapi\Client\Infrastructure\RequestExecutor\RequestExecutorFactory;
 use Smsapi\Client\Service\SmsapiComService;
 use Smsapi\Client\Service\SmsapiComHttpService;
@@ -21,22 +24,24 @@ class SmsapiHttpClient implements SmsapiClient
 {
     use LoggerAwareTrait;
 
+    private $httpClient;
+    private $requestFactory;
+    private $streamFactory;
+
     private $smsapiPlUri = 'https://api.smsapi.pl';
     private $smsapiComUri = 'https://api.smsapi.com';
-    private $proxy = '';
     private $dataFactoryProvider;
 
-    public function __construct()
-    {
+    public function __construct(
+        ClientInterface $httpClient,
+        RequestFactoryInterface $requestFactory,
+        StreamFactoryInterface $streamFactory
+    ) {
+        $this->httpClient = $httpClient;
+        $this->requestFactory = $requestFactory;
         $this->dataFactoryProvider = new DataFactoryProvider();
         $this->logger = new NullLogger();
-    }
-
-    public function setProxy(string $proxy): SmsapiClient
-    {
-        $this->proxy = $proxy;
-
-        return $this;
+        $this->streamFactory = $streamFactory;
     }
 
     public function smsapiPlService(string $apiToken): SmsapiPlService
@@ -47,6 +52,7 @@ class SmsapiHttpClient implements SmsapiClient
     public function smsapiPlServiceWithUri(string $apiToken, string $uri): SmsapiPlService
     {
         return new SmsapiPlHttpService(
+            $this->httpClient,
             $this->createRequestExecutorFactory($apiToken, $uri),
             $this->dataFactoryProvider
         );
@@ -60,6 +66,7 @@ class SmsapiHttpClient implements SmsapiClient
     public function smsapiComServiceWithUri(string $apiToken, string $uri): SmsapiComService
     {
         return new SmsapiComHttpService(
+            $this->httpClient,
             $this->createRequestExecutorFactory($apiToken, $uri),
             $this->dataFactoryProvider
         );
@@ -67,8 +74,8 @@ class SmsapiHttpClient implements SmsapiClient
 
     private function createRequestExecutorFactory(string $apiToken, string $uri): RequestExecutorFactory
     {
-        $guzzleClientFactory = new GuzzleClientFactory($apiToken, $uri, $this->proxy);
-        $requestExecutorFactory = new RequestExecutorFactory($guzzleClientFactory);
+        $httpClientFactory = new HttpClientFactory($apiToken, $uri);
+        $requestExecutorFactory = new RequestExecutorFactory($httpClientFactory, $this->requestFactory, $this->streamFactory);
         $requestExecutorFactory->setLogger($this->logger);
 
         return $requestExecutorFactory;