+++ tests/Helper/HttpClient/HttpClientMock.php
@@ -1,0 +1,51 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Smsapi\Client\Tests\Helper\HttpClient;
+
+use GuzzleHttp\Client;
+use GuzzleHttp\Exception\GuzzleException;
+use GuzzleHttp\Exception\RequestException as GuzzleRequestException;
+use GuzzleHttp\Exception\TransferException as GuzzleTransferException;
+use GuzzleHttp\Handler\MockHandler;
+use GuzzleHttp\HandlerStack;
+use GuzzleHttp\Psr7\Response;
+use Psr\Http\Client\ClientInterface;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
+use Smsapi\Client\Tests\Helper\HttpClient\Exception\ClientException;
+use Smsapi\Client\Tests\Helper\HttpClient\Exception\NetworkException;
+use Smsapi\Client\Tests\Helper\HttpClient\Exception\RequestException;
+
+class HttpClientMock implements ClientInterface
+{
+    private $mockHandler;
+
+    public function __construct()
+    {
+        $this->mockHandler = new MockHandler();
+    }
+
+    public function sendRequest(RequestInterface $request): ResponseInterface
+    {
+        $guzzleClient = new Client([
+            'handler' => HandlerStack::create($this->mockHandler)
+        ]);
+
+        try {
+            return $guzzleClient->send($request);
+        } catch (GuzzleRequestException $e) {
+            throw RequestException::create($request, $e);
+        } catch (GuzzleTransferException $e) {
+            throw NetworkException::create($request, $e);
+        } catch (GuzzleException $e) {
+            throw ClientException::create($request, $e);
+        }
+    }
+
+    public function mockResponse(int $statusCode, string $body)
+    {
+        $this->mockHandler->append(new Response($statusCode, [], $body));
+    }
+}